requirements.txt
streamlit
OPENAI
python-dotenv
langchain_openai
google.generativeai